create page refactor - in progress
