+++ frontend/src/config/tags.ts
@@ -1,0 +1,53 @@
+export const tags = [
+  'innovation',
+  'creativity',
+  'discovery',
+  'experiment',
+  'prototyping',
+  'tinkering',
+  'build',
+  'invention',
+  'crafting',
+  'design',
+  'making',
+  'exploration',
+  'challenge',
+  'hands on',
+  'brainstorming',
+  'problem-solving',
+  'iteration',
+  'conceptualization',
+  'imagination',
+  'construction',
+  'modeling',
+  'improvement',
+  'collaboration',
+  'curiosity',
+  'ideation',
+  'solution-driven',
+  'blueprinting',
+  'fabrication',
+  'customization',
+  'learning by doing',
+  'engineering',
+  'testing',
+  'creativity boost',
+  'innovation lab',
+  'craftsmanship',
+  'handcrafted',
+  'skill development',
+  'curiosity-driven',
+  'articulation',
+  'prototype testing',
+  'knowledge expansion',
+  'digital fabrication',
+  'adaptation',
+  'resourcefulness',
+  'open-ended projects',
+  'technical skills',
+  'exploratory learning',
+  'hands-on practice',
+  'skill mastery',
+  'manual work',
+  'analytical thinking',
+];
